backend/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
SQLAlchemy==2.0.32
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
flask-cors==4.0.0
twilio>=9.0.0

